+++ src/configuration/data/LiquibaseConfigurationData.ts
@@ -212,10 +212,15 @@ export class LiquibaseConfigurationData {
    * @param pBuildDriverPath - function to build the driver path.
    * This function can not be implemented in the data classes,
    * because building the path requires the `path` module and this is not allowed in classes that are used in a webview.
+   * @param pDisguisePassword - if the password should not be displayed as plain text, but as `***`. These should be used if you are in a preview.
+   * If no value is given, then the password will be set as plain text
    * @returns the created properties file as a string
    */
-  generateProperties(pBuildDriverPath: (pDriver: Driver) => string | undefined): string {
-    const propertiesEditor = this.generatePropertiesEditor(pBuildDriverPath);
+  generateProperties(pBuildDriverPath: (pDriver: Driver) => string | undefined, pDisguisePassword?: boolean): string {
+    const propertiesEditor = this.generatePropertiesEditor(
+      typeof pDisguisePassword === "undefined" ? false : pDisguisePassword,
+      pBuildDriverPath
+    );
     // replace all escaped colons with unescaped.
     // There is no way to automatically escape them during creation
     // TODO maybe more escapes are needed with unescapeContent
@@ -224,12 +229,16 @@ export class LiquibaseConfigurationData {
 
   /**
    * Creates the properties editor for the given configuration.
-   *  @param pBuildDriverPath - function to build the driver path.
+   * @param pDisguisePassword - if the password should not be displayed as plain text, but as `***`. These should be used if you are in a preview.
+   * @param pBuildDriverPath - function to build the driver path.
    * This function can not be implemented in the data classes,
    * because building the path requires the `path` module and this is not allowed in classes that are used in a webview.
    * @returns the created properties
    */
-  private generatePropertiesEditor(pBuildDriverPath: (pDriver: Driver) => string | undefined): PropertiesEditor {
+  private generatePropertiesEditor(
+    pDisguisePassword: boolean,
+    pBuildDriverPath: (pDriver: Driver) => string | undefined
+  ): PropertiesEditor {
     // Build the properties
     const properties: PropertiesEditor = new PropertiesEditor("");
 
@@ -242,13 +251,23 @@ export class LiquibaseConfigurationData {
     classpathElements.push(this.liquibaseSettings.liquibaseDirectoryInProject);
 
     if (this.databaseConnection.hasData()) {
-      const result = this.databaseConnection.writeDataForConnection(properties, false, pBuildDriverPath);
+      const result = this.databaseConnection.writeDataForConnection(
+        properties,
+        false,
+        pBuildDriverPath,
+        pDisguisePassword
+      );
       result && classpathElements.push(result);
     }
 
     // and the reference properties
     if (this.referenceDatabaseConnection && this.referenceDatabaseConnection.hasData()) {
-      const result = this.referenceDatabaseConnection.writeDataForConnection(properties, true, pBuildDriverPath);
+      const result = this.referenceDatabaseConnection.writeDataForConnection(
+        properties,
+        true,
+        pBuildDriverPath,
+        pDisguisePassword
+      );
       result && classpathElements.push(result);
     }